Moved the SyncEnv into core.py
